requirements.txt
altgraph==0.16.1
beautifulsoup4==4.6.3
bs4==0.0.1
cachetools==2.1.0
cairocffi==0.9.0
CairoSVG==1.0.22
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.0.3
dateparser==0.7.0
dicttoxml==1.3.1
diff-match-patch==20121119
Django==1.11.2
django-admin-report==0.4
django-ckeditor==5.2.2
django-crispy-forms==1.6.1
django-daterange-filter==1.3.0
django-extensions==1.7.9
django-filter==1.0.4
django-geojson==2.10.0
django-import-export==0.4.3
django-leaflet==0.22.0
django-localflavor==1.5.1
django-mathfilters==0.4.0
django-multiupload==0.5.2
django-registration==2.2
django-report-utils==0.3.1
django-rest-swagger==2.1.2
django-simple-history==1.9.0
django-tables2==1.9.0
django-uuidfield==0.5.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.1
django-wysiwyg==0.8.0
djangorestframework==3.6.3
et-xmlfile==1.0.1
factory-boy==2.10.0
Faker==0.8.12
funcsigs==1.0.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.10.3
idna==2.5
ipaddress==1.0.22
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
ldap==1.0.2
ldap3==2.5.1
lxml==4.2.5
macholib==1.8
MarkupSafe==1.0
mock==2.0.0
MySQL-python==1.2.5
oauthlib==2.1.0
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.4
pbr==4.3.0
Pillow==2.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
Pyphen==0.9.5
python-dateutil==2.6.0
python-ldap==3.0.0
python-logstash==0.4.6
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
regex==2018.8.29
reportlab==3.4.0
requests==2.18.1
requests-oauthlib==0.8.0
rsa==4.0
selenium==3.14.1
simplejson==3.11.1
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.7.0
sqlparse==0.2.4
tablib==0.12.1
text-unidecode==1.2
tinycss2==0.6.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.21.1
WeasyPrint==0.39
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
